requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
python-multipart>=0.0.5
pydantic>=1.8.0
fastapi-cache2>=0.2.1
redis>=4.5.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0 